requirement.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
psycopg2-binary==2.9.5
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
soupsieve==2.4
SQLAlchemy==2.0.7
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
